+++ lib/ruby-measurement.rb
@@ -1,3 +1,4 @@
 require 'ruby-measurement/version'
+require 'ruby-measurement/core_ext'
 require 'ruby-measurement/measurement'
 require 'ruby-measurement/definitions'