+++ lib/ruby-measurement/core_ext.rb
@@ -1,0 +1,3 @@
+require 'ruby-measurement/core_ext/numeric'
+require 'ruby-measurement/core_ext/string'
+require 'ruby-measurement/core_ext/symbol'