+++ lib/ruby-measurement/core_ext/numeric.rb
@@ -1,0 +1,7 @@
+require 'ruby-measurement/measurement'
+
+class Numeric
+  def to_measurement
+    Measurement.new(self)
+  end
+end