+++ lib/ruby-measurement/core_ext/symbol.rb
@@ -1,0 +1,7 @@
+require 'ruby-measurement/unit'
+
+class Symbol
+  def to_unit
+    Measurement::Unit[self] or raise ArgumentError, "Invalid unit: #{self}"
+  end
+end